requirements.txt
torch
torchvision
opencv-python
matplotlib
Pillow
numpy==1.26 # numpy 2.0 not compatible at the moment
scipy
scikit-learn
faiss-gpu
tqdm
tifffile
argparse
PyYAML
pandas
timm==0.6.13
packaging
